Create PSv1.0.pl
This is a basic ping script that will prompt for a range of ips and the SNMPv2 community. It will then send a single ping to all the ip addresses in the range specified. Once the pings are finished it will SNMPwalk -v 2c those pings that reported alive. After it finishes with SNMPv2 polling it will ask the user if they want to try SNMPv3. It prompts the user for the user security model level, user name security view, community, encryption passphrase and level of encryption. I wrote this script to assist in network discovery but I ended up using it to mainly perform SNMPv3 walks because it has all the different commands structures built into it.